Add logging to rules operations in supervdsmServer
Previously when udev rules were created and removed, we had no logging,
so trying to debug issues with udev rules was very hard. This patch log
creation and removal of udev rules.

Although the whole methods were logged using logDecorator the creation
and removal of each rule was not logged independently

Change-Id: Ib1d8ede0cbe0790bdfcd60ed833b3d9c54f31eba
Signed-off-by: Xavi Francisco <xfrancis@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26121
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>