caps: Collect kdump status
Adds kdump configuration status to caps module. The status can contain:

  -1 - UNKNOWN (it's set if some error appears during status check)
   0 - DISABLED (kdump is not loaded or fence_kdump is not configured)
   1 - ENABLED (kdump is loaded properly and fence_kdump is configured)

The status will be reported as part of getCapabilities response in key
'kdumpStatus'.

The value will be displayed in webadmin in General tab inside Host view
and it will also be used in kdump detection part in fencing flow.

Change-Id: I68d7a2a24fdaad74255004af0f327197eaee65f2
Bug-Url: https://bugzilla.redhat.com/1079821
Signed-off-by: Martin Perina <mperina@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25926
Reviewed-by: Barak Azulay <bazulay@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>