init: Normalize shutdown timeout
On sysvinit, vdsm was killed 2 seconds after receiving a SIGTERM.
Currently clientIF check if shutdown is started only once every 3
seconds. This make it impossible to write profile data when vdsm is
terminated. On systemd, since no timeout was defined, the default 90
seconds timeout was used. On upstart, the default 5 seconds was used.

This patch normalizes the shutdown timeout to 10 seconds on all
platforms. This seems to be enough for saving profile data and
performing clean shutdown.

Change-Id: I5434caff5bc95e2a76345191dec9ce4e33070a57
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26111
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29674