core: fixing attribute error on fenceSpmStorage
In change I75493d1db60e51cccd5231b516f963c970d24c99 the pool metadata
operations were moved to the pool backend, while the call to
invalidateMetadata in fenceSpmStorage was left as it was, leading to an
attribute error.

This patch fixes it by executing the operation on the pool backend object
rather than on the pool, as the metadata operation are implemented only in
StoragePoolDiskBackend - there's also a check for the backend type before
executing it.
Though it's not the nicest way for doing so, it's the pattern already used in
hsm.py,therefore in the meanwhile it can be fixed the same way.

Change-Id: I5ae4c371e2dc97686621ea16b8cef971183a6ab2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1082365
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26358
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27341
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>