pad memory volume only when the storage domain is file based
Memory volume should be padded only when the storage domain in which the
memory volume resides is file-based device. We used to check the type of
the storage pool, assuming that all the domains within the pool are from
the same type. This assumption is not true anymore, as we can have
different types of storage domains in the same storage pool, in case of
shared storage. So from now on, we'll check the type of the storage
domain in which the memory volume is in, and only if it is filed-based
device we'll pad the volume.

Change-Id: I0cebdbba2cc763963b64dcbc142f979802f18c16
Bug-Url: https://bugzilla.redhat.com/1082941
Signed-off-by: Arik Hadas <ahadas@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26407
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26594
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>