recovery: clean the VM recovery procedure
This tiny patch cleans up and optimizes the VM recovery detection.
While looping on the VMs returned by libvirt, we must check each one
of them to know if it has to be handled by VDSM.
This check, however, considers only the properties of the VM under exam,
so we can do this in one go instead of using two loops
(actually an explicit loop and one implicit in a list comprehension)

This patch is not for performance but to to make the code a tiny bit
cleaner and to fix this message:

./vdsm/clientIF.py:493: list comprehension redefines 'vm' from line 483

emitted by some versions of pyflakes.

Change-Id: Idf09400276d9fcf24a0ad6673ee1d12f19a92195
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26426
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>