agent: number-of-cpus reporting support
This patch implements handling of the new message 'number-of-cpus'

The 'count' argument of the message sent by the guest agent is an integer
and represents the number of CPU cores are visible on the guest OS side.
If the value is -1, the reporting is not supported or implemented for the
guest OS.

The value is exposed to the engine via RunningVmStats structure as
'guestCPUCount' value.

Change-Id: I37a4ac7ccf16a7d008f737c16fd1c60293ff2272
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1063281
Guest-Agent-API-Version: 1
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26530
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27035
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>