tests: virt: retry to connect to VDSM at startup
This patch makes the functional tests retry to connect
to the VDSM instance before to give up and raise an error.

The purpose is let tests to actively wait for VDSM
instance to be up instead of blindly wait some time and
hope for the best.

This will help us to drop the 10s sleep we currently
have in our virt functional tests job.

Change-Id: Ic3a6e9e484d7bf0c506b804131b9e958deff5692
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26638
Reviewed-by: Martin Polednik <mpoledni@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>