ethtool_opts: Add hook for applying network custom property
ethtool_options is a hook that takes as input a network 'custom'
property named 'ethtool_opts' that has as its content the
arguments for the ethtool command line utility.

This hook detects whether those arguments are valid in the sense
of applying to nics that belong to the network hierarchy for which
the custom property has been set.

As an example of the expected content:
    '--offload em2 rx on --offload em1 tx on'
which would apply for a bond that is set over em1 and em2.

Change-Id: I41d89e00002e2b011bc1dc2825276d14bc99f577
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26694
Reviewed-by: Dan Kenigsberg <danken@redhat.com>