log: Change vdsm log file to enforce user/group
This commit adds a new log handler that extends WatchedFileHandler,
This handler checks current user and group and checks with the expected
values, to make sure logs are created with correct permissions. In case of bad
access RuntimeError is thrown.

This patch also updates logger.conf to use this handler.

Change-Id: I0a4d7212cb311b22e4fb60ffdc45163a496a74d6
Signed-off-by: Dima Kuznetsov <dkuznets@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26728
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>