core: introducing uploadImageToStream
Adding uploadImageToStream  verb which allows to retrieve image data from
vdsm to the requester over HTTP.
GET requests arriving to the server  to default paths that we use
today for request handling ('/', '/RPC2') will be treated
as retrieve requests.
The retrieve operation  itself is being executed within a task, that's
needed to indicate that there's an operation executed by
the host.

Change-Id: If2df4d3a16f39bf80281d7669ed31fd8369bada5
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26741
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>