jsonrpc: Stomp support
The mini broker turned up to be much more work than I expected.
This is an intermediate solution. VDSM will accept broker like
message but will ignore most of them and not enforce policy.

This is so that we can get the engine talking in the correct way and
work on the mini broker stress free. This also means that we can release
VDSM without the mini broker using a message format that is future
ready.

It is expected from the engine to send the CONNECT and SUBSCRIBE frames
even though VDSM doesn't yet enforce them because future VDSM's probably
will.

Change-Id: I22bcae1e150dea7bc7d9fecefb6847c48bfe8949
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26750
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>