task: support scheduleJob without persistency
At the moment we are not using any hsm task (the only one is spmStart
and it's not relying on _hsmSchedule).
This patch adds the support in taskManager to scheduleJob without any
persistency, removes the reload of dumped tasks from storageRefresh,
and finally redefines _hsmSchedule to not use persistency.
The unused method _hsmSchedule is not removed because it will be used
by spmStart in a following patch.

Change-Id: I95b3063048b910dc5994d0a264bf4791e98fef51
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26771
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>