LiveMerge: Add support for deleting merged volumes
Once a live merge has completed the merged volumes that have previously
been unlinked from the chain need to be deleted.  We can use the
existing deleteVolume verb with one minor change for block volumes.
When deleting a live-merged volume, its former child may have a stale LV
tag pointing back to this volume.  This occurs because an HSM host
resolves the block job completion event and can only update the parent
UUID in the storage domain's metadata LV (not in the LV Tags).  In this
special case, we sync the LV tag of the child to match the metadata LV
prior to validating the delete operation.

Also, don't validate deletes in hsm.py since we already perform
validation in the volume-specific implementations.

Change-Id: Icf43f65f984d7f76622454dc8877d56ef8dedb8f
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26768
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>