vm: Fix guest channel symlink creation handling
In case of a restart of VDSM the symlinks and sockets won't get cleaned up.
Commit 4a7af02 partially fixed the issue, however it did not consider the
case of a crashing VDSM. In case of a crashing or unclean stop (kill)
of VDSM the symlinks are left behind and the recovery would fail.

This patch handles these cases and adds a bit more of documentation to the
code to improve the clarity of the cases it handles.

Change-Id: I24e5d8e7b7e338c683d4064585539444044bb77a
Bug-Url: https://bugzilla.redhat.com/1082986
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26882
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27183
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>