virt: graphdev: add the GraphicsDevice class
this patch introduces the Graphics Device class to be used
to define the Vm display properties.
No more than one Graphics Device is supported.

If VDSM receives a Graphic Device definition, it uses that;
otherwise, tries to fallback to display* parameters received as input.

VDSM will continue to supply the legacy display* parameters
in stats output and in the conf dictionary (VmDefinition)
to preserve the backward compatibility. This includes migrations,
so a VM created with a single new style graphic device can be
migrated to old VDSMs without breaking cluster compatibility.

Change-Id: I82e4414aa3efe7c756ba8ef9b7b47e591613a717
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26895
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>