vm: floppy: remove special 'readonly' handling
floppy devices used to have an explicit extra check
on the image permissions to determine if the image
had to have the 'readonly' attribute set.

Engine is already sending this information, and this
check is already done previously in vm.Drive.getXML()

Moreover, this check is not only redundant, but it is
one of the legacy remaining places vdsm can get stuck
if the ISO domain becomes inaccessible when a VM is
started, because no out-of-process I/O is used.

To address those issues, this patch always mark the
floppy image read-only, and removes the explicit
additional check and all the code which was used
only by it.

This patch purposefully undoes what was done to fix
https://bugzilla.redhat.com/show_bug.cgi?id=674357
because nowadays there is no longer use for floppy
devices outside sysprep, but this is already out of
direct user control.

Change-Id: Ie9183104d3739f2d4a99712e7851fc3172b82621
Bug-Url: https://bugzilla.redhat.com/755528
Relates-To: https://bugzilla.redhat.com/674357
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26937
Reviewed-by: Dan Kenigsberg <danken@redhat.com>