unified_persistence: Fix defaultRoute type in upgrade script
Despite the fact defaultRoute is defined by vdsm API to be a boolean
field, the upgrade to unified persistence script was retrieving it
as a string. This had the effect that when the ifcfg configurator
would convert from boolean to 'yes'/'no', both 'True' and 'False',
being strings would evaluate as True.

Change-Id: I0f468bed70fcb3e2af46944f10ef47eecb1c0e94
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27021
Reviewed-by: Dan Kenigsberg <danken@redhat.com>