ifcfg: fix stp processing
Some users rely on passing STP=on/off for setting a bridge's stp.
This worked by passing the option through all the way to the ifcfg
configurator and have it just copy the key-value to the ifcfg files.
The problem is that STP was not being popped out of opts, so 'STP'
would be written to the bridge port devices as well.

This patch normalizes stp into the Bridge model and takes it out of
the passthrough.

Change-Id: I6a1a37485ae40b8c9f2037c671daa3e4c6ac3f90
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27023
Reviewed-by: Dan Kenigsberg <danken@redhat.com>