caps: Repair collect numa information
This patch repairs the memory issue about collect numa information
on a non-numa support machine.
If there is only one cell in the result of libvirt's getCapabilities
api, we consider it as non-numa support machine. In this situation,
it will be issue an error when call libvirt's getMemoryStats api.
Instead, we get the memory stats through /proc/meminfo.

Change-Id: Ifb70231c81e75d29107f4ea89367aa33d8e85bad
Bug-Url: https://bugzilla.redhat.com/1069303
Signed-off-by: Bruce Shi <xiao-lei.shi@hp.com>
Reviewed-on: http://gerrit.ovirt.org/27090
Reviewed-by: Dan Kenigsberg <danken@redhat.com>