sampling: Collect numa related statistics
This patch modifies the HostStatsThread in sampling module to add
function of collect host numa related statistics data.
It will add two keys in the rpc response of getVdsStats whose value
are the host's numa related statistics information.
Key cpuStatistics, the value is a map, each item follows this format:
{'cpuCoreId':{'nodeIndex':int, 'cpuIdle':'str', 'cpuSys':'str',
              'cpuUser':'str'}}
cpuCoreId      - The id of this cpu core
nodeIndex      - The numa node index that this cpu core belongs
cpuIdle        - The idle percentage of this cpu core
cpuUser        - The user consumed percentage of this cpu core
cpuSys         - The system consumed percentage of this cpu core
Key numaNodeMemFree, the value is a map, each item follows this
format:
{'nodeIndex':{'memFree':str, 'memPercent':int}}
nodeIndex      - The index of numa node
memFree        - The free memory of this numa node in MB
memPercent     - The used memory percentage of this numa node

Change-Id: Ib8d2edd74b8b5d5d3d3e508afc3f675037a86597
Bug-Url: https://bugzilla.redhat.com/1069303
Signed-off-by: Bruce Shi <xiao-lei.shi@hp.com>
Reviewed-on: http://gerrit.ovirt.org/26876
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>