multipath: Rescan also FC devices
When rescanning devices, we used to update ISCSI devices, and then run
multipath to update device mapping. However, we never rescanned FC
devices.

We used to believe that "multipath -r" was triggering a rescan of FC
devices and solving this issue. This may have been true in older
versions but certainly does not work with current RHEL release, and was
removed in the previous patch.

This patch adds the missing FC rescan using the issue_lip sysfs api,
recommended in the RHEL Storage Administration Guide.

Change-Id: I7699504f9771232ee0b880f9c83a51fd5b90f40e
Bug-Url: https://bugzilla.redhat.com/1121998
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27122
Tested-by: Elad Ben Aharon <eladba1990@gmail.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Xavi Francisco <xfrancis@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30984
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>