virt: Capture migration start time after the semaphore was accquired
Previously the migration start time was captured at the start of the
MigrationSourceThread which leads to aborting migrations when the VMs
had too long to wait for the accquisition of the migration semaphore.

This patch captures the start time in the moment the migration really
begins. Additionally this patch logs the time the migration took to
accquire the migration semaphore to improve debugging of the logs.

Change-Id: I209023f5f4a3ccee5db4d43b17627b09ca55bef8
Bug-Url: https://bugzilla.redhat.com/1090109
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27135
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>