vm: drop wasteful time.sleep() in releaseVm
After inspection of the libvirt code, turns out
the time.sleep(30) we currently have in the releaseVm
flow is redundant and wasteful.

Libvirt (1.2.x) already has hardcoded timeout to handle
the QEMU process termination, and this patch leverages that.

Before this patch, the sequence of events triggered
by the dom.destroy() call in the worst case scenario
(QEMU not responding) is:
- send SIGTERM (clean shutdown request)
- wait 10s (libvirt, hardcoded)
- (fails)
- log error (libvirt, vdsm)
- wait 30s (vdsm)
- send SIGTERM (clean shutdown request)
- wait 10s (libvirt)
- (fails)
- log error (libvirt)
- send SIGKILL (forced shutdown request)
- wait 5s( libvirt, hardcoded)
total time: 10+30+10+5 = 55s (all hardcoded!)

With this patch, the flow becomes:
- send SIGTERM (clean shutdown request)
- wait 10s (libvirt, hardcoded)
- (fails)
- log error (libvirt, vdsm)
- send SIGTERM (clean shutdown request)
- wait 10s (libvirt)
- (fails)
- log error (libvirt)
- send SIGKILL (forced shutdown request)
- wait 5s( libvirt, hardcoded)
total time: 10+10+5 = 25s (no values hardcoded in VDSM)

Change-Id: Ia25794217e63dd4c755d9c141a2e941e4baa6fd3
Bug-Url: https://bugzilla.redhat.com/1091389
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27175
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>