mutipath: Remove unneeded and dangerous -r parameter
Since commit dbf2089488 (Jul 9 2013) multipath call was change to use
the -r flag, forcing a reload of the device map. This was tested to fix
a case where new lun is created on the storage server, while a host was
connected, and the new device is not available when issuing the
getDeviceList command. According to a comment on gerrit, the change was
tested for ISCSI and FC storage types, but there is no documentation of
the testing procedure. The related bug was verified, but has no
information about how it was verified.

We have two related bugs:

- Bug 1078879 tell us that invoking multipath with the -r flag sometimes
  triggers a segfault in the multipathd daemon. In the bug, multipath
  developer suggests that as long as multipathd daemon is running,
  there is no need to invoke multipath to detect new devices, and
  "multipath -r really isn't useful for much of anything".

- Bug 1071654 tell us that devices rescanning is broken on FC storage
  domains (although the -r flag is used). I reproduced this bug using
  storage QE FC server.

This patch removes the -r flag. To be on the safe side, I left the
multipath call as it was since the first multipath commit in 2009. We
will work with kernel and multipath developers further on removing this
call if it is indeed unneeded.

Bug-Url: https://bugzilla.redhat.com/1078879
Relates-to: https://bugzilla.redhat.com/1071654
Relates-to: http://gerrit.ovirt.org/17263
Change-Id: I880ab5343df3e0030638901e188320b20570747d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27242
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>