sp: update domain links on state change
In some situations (e.g. after a reconstructMaster command) the
domain links may be missing and it's impossible to recreate them
on connectStoragePool or refreshStoragePool since the storage may
be unreachable. The best we can do is recreate them as soon as
the domains are visible again.

This patch has been submitted to the ovirt-3.4 branch only because
it's not the long-term solution that instead should be delivered
by BZ#1093924.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1086210
Relates-To: https://bugzilla.redhat.com/show_bug.cgi?id=1093924
Change-Id: I7ac20e4b515472b24c35b2cccd2ad3dc98b3574c
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27466
Reviewed-by: Dan Kenigsberg <danken@redhat.com>