profile: Add couple of tests
We plan to modify the profiler so it will be useable for the functional
tests, where we like to have one profile per test, instead of the
application wide profile.

This patch adds couple of tests to make it easier to modify the profiler
without breaking anything.

The profile filename and format are now module private constants, so we
can easily override them during the tests.

Change-Id: I0dedc21993972ec25881f46f0b365445ed494aac
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27328
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>