net_tc: add support for mirroring all (not just IPv4) traffic
Up until now the filter was only acting on packets with IPv4 headers
because of the 'protocol ip' part of the tc filter. This makes the
mirroring protocol agnostic.
Caveat: this patch takes place only on networks defined after its
application. It does not attempt to upgrade existing filters.

Change-Id: I9de381e67fae4cb79fe41dbd9d1b60b72eb84de4
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>
Reviewed-on: https://gerrit.ovirt.org/27388
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>