stats: avoid log if disk stats are known missing
There are some known cases on which the disk stats cannot
be available, for example if recovery is in progress.

Those cases are already handled by the current code,
but a message is logged anyway, even if it is a known
and benign case of  expected missing sample.

This patch add explicit handling for the benign missing
disk stats case, to avoid to produce log noise when it is
expected to not have the disk stats.

Change-Id: Ia4dfd5f51fa3ed8cba172dbff0ae23714cabaf67
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27414
Reviewed-by: Dan Kenigsberg <danken@redhat.com>