stats: avoid log if disk latency is known missing
There are some known cases on which the disk latency stats
are not available, for example if recovery is in progress.

Those cases are already handled by the current code,
but a message is logged anyway, even if it is a known
and benign case of  expected missing sample.

This patch add explicit handling for the benign missing disk
latency stats case, to avoid to produce log noise when it is
expected to not have the disk latency stats.

Change-Id: I1119ec792ff8aba6e5340efd0c1b7e2ee89a296f
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27415
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>