stats: avoid log if cpu stats are known missing
There are some known cases on which the cpu stats cannot
be available, for example if not enough samples are yet
collected.

Those cases are already handled by the current code,
but a message is logged anyway, even if it is a known
and benign case of  expected missing sample.

This patch add explicit handling for the benign missing
cpu stats case, to avoid to produce log noise when it is
expected to not have the cpu stats.

Change-Id: I62eff2e04a2922fb683d3c1d78605990399c7b76
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27416
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>