vm: add a method to update the devices config
As part of all creation flows except the recovery one,
a Vm needs to amend the configuration of the devices
and update its own configuration accordingly.

This patch extracts the code which updates the
Vm.conf devices data and puts it into a separate
method to make the code a bit more clear.

Change-Id: I04dab67c60c51a381b61410bac8c87f5e29809d1
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27443
Reviewed-by: Dan Kenigsberg <danken@redhat.com>