tests: Add symlink mount test
Some filesystems like gfs2 resolves symlink (as /dev/mapper/name or
/dev/vg/lv), and the real device will appear in /proc/mounts instead of
the original device symlink specified on the command line. This adds a
test for this case.

This test use a symlink to a backing image mounted using the loop
device. This is a simple way to simulate the same issue when using gfs2.

Change-Id: I9f3d6333921505846f345d015907e5f6174af4f8
Bug-Url: https://bugzilla.redhat.com/888711
Signed-off-by: Enrico Tagliavini <enrico.tagliavini@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/27514
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Yoav Kleinberger <ykleinbe@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>