hsm: fix setStoragePoolDescription execution
In commit ab6014b the pool metadata operations were moved to the pool
backend, while the call to setStoragePoolDescription was left as it was,
leading to an attribute error.

StoragePoolMemoryBackend does not maintain any information about the
pool so setStoragePoolDescription it is not supported.
It is safe to ignore the setStoragePoolDescription request because in
StoragePoolMemoryBackend there's no place where to persist the value
and engine never tries to read this information.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1096212
Change-Id: Ic85ecd5310d15947631224a21f962526eec91180
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27540
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>