virt: graphdev: support multiple graphics devices
This patch allows a Vm to have more than one Graphics Devices,
so a we can now have multiple consoles (e.g. SPICE and VNC
at the same time).

With this feature, a user can switch clients without restarting that VM.
Another possible use case is the regular users using SPICE for
performance and the administrator using VNC for portability.

Bug-Url: https://bugzilla.redhat.com/976044
Change-Id: I5be348b342359d42c878937dca27454fe206a35a
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27215
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>