vm: hyperv: initial windows hyperv support
This patch adds explicit support for hyperv optimizations.
The optimizations are both for stability and for performance.
Engine, or any other client, can enable them by supplying a new
optional parameter 'hyperVenable' in the Vm parameters at
creation time.

As default, the new settings are disabled for backward compatibility.

The parameters are hardcoded and not externally configurable
because they are not supposed to be changed very often, if
changed at all; moreover, we implement the optimal recommended
settings.

Change-Id: I28ea1d5adeda07798255484209e1a1d92c2c2bc5
Bug-Url: https://bugzilla.redhat.com/1110305
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27619
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/30254
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>