domainMonitor: Extract domain monitoring methods
The _domainMonitor method was extremely long and was doing too much.
This patch breaks it to multiple small methods to make it easier to
understand and modify.

One of the planned changes is to allow _domainMonitor to abort in the
middle if a monitor is stopped. Implementing this in the original method
would only make it worse. To make it easy, operation that may take lot
of time were extracted into separate methods.

This patch does not change the behavior of the monitor, only move code
around.

Change-Id: I87ac6a82e560bc4360a3bc3f6f6fd94623678cc2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27714
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>